Extract PayPalNativeView.
